x86_64 JIT support
svn: r2621

+++ src/mzscheme/src/lightning/i386/funcs.h
@@ -53,11 +53,11 @@ jit_flush_code(void *dest, void *end)
      jit_flush_code as an mprotect.  */
 #ifdef __linux__
   static unsigned long prev_page = 0, prev_length = 0;
-  int page, length;
+  long page, length;
 #ifdef PAGESIZE
-  const int page_size = PAGESIZE;
+  const long page_size = PAGESIZE;
 #else
-  static int page_size = -1;
+  static long page_size = -1;
   if (page_size == -1)
     page_size = sysconf (_SC_PAGESIZE);
 #endif
